Remove the `eval` indirection
There is no need to eval the backend because we can just import it.
